+++ src/result.rs
@@ -17,8 +17,17 @@ pub enum Error {
     /// IO error that might happen during command / process execution.
     #[error("IO error: {0}")]
     IoError(io::Error),
+    /// Process was interrupted by user (`Ctrl + C`).
+    #[error("Interrupted.")]
+    Interrupted,
+    /// Process was killed because it couldn't exit gracefully.
+    #[error("Killed [pid: {pid}].")]
+    Killed {
+        /// Process identifier.
+        pid: u32,
+    },
     /// Error raised when a process exits with a non-zero exit code.
-    #[error("Process exited with non-zero code: {:#?}. Output: {:#?}", .code, .output)]
+    #[error("Process exited with non-zero code: {code:?}. Output: {output:#?}")]
     NonZeroExitCode {
         /// Exit code of a process. Might be absent on Unix systems when a process was terminated by a signal.
         code: Option<i32>,
@@ -33,7 +42,7 @@ pub enum Error {
     /// When a process manager failed to kill hanged child process, there is a zombie process left hanging around.
     /// This error provides details, such as process id and an error, so user could handle cleaning manually.
     #[cfg(unix)]
-    #[error("Process with pid {pid} hanged and we were unable to kill it. Error: {err}", pid = .pid, err = .err)]
+    #[error("Process with pid {pid} hanged and we were unable to kill it. Error: {err}")]
     Zombie {
         /// Process id of the hanged process.
         pid: u32,
@@ -43,7 +52,7 @@ pub enum Error {
     /// When a process manager failed to kill hanged child process, there is a zombie process left hanging around.
     /// This error provides details, such as process id and an error, so user could handle cleaning manually.
     #[cfg(windows)]
-    #[error("Process with pid {pid} hanged and we were unable to kill it. Error: {err}", pid = .pid, err = .err)]
+    #[error("Process with pid {pid} hanged and we were unable to kill it. Error: {err}")]
     Zombie {
         /// Process id of the hanged process.
         pid: u32,